Correction bug edit
